requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.8.2
django-celery==3.1.17
djangorestframework==3.4.0
kombu==3.0.35
pytz==2016.6.1
redis==2.10.5
